+++ components/common/src/main/java/com/hotels/styx/common/format/SanitisedHttpMessageFormatter.java
@@ -96,7 +96,7 @@ public class SanitisedHttpMessageFormatter implements HttpMessageFormatter {
 
     @Override
     public Throwable wrap(Throwable t) {
-        return t == null ? null : SanitisingThrowableFactory.instance().create(t, sanitisedHttpHeaderFormatter);
+        return t == null ? null : ThrowableSanitiser.instance().sanitise(t, sanitisedHttpHeaderFormatter);
     }
 
     private String formatNettyRequest(io.netty.handler.codec.http.HttpRequest request) {
