+++ pom.xml
@@ -118,7 +118,7 @@
     <reactive-streams.version>1.0.2</reactive-streams.version>
     <reactor.version>3.3.0.RELEASE</reactor.version>
     <pcollections.version>3.0.3</pcollections.version>
-    <cglib.version>3.3.0</cglib.version>
+    <bytebuddy.version>1.10.6</bytebuddy.version>
 
     <!--Kotlin -->
     <kotlin.version>1.3.21</kotlin.version>
@@ -336,7 +336,7 @@
       <dependency>
         <groupId>net.bytebuddy</groupId>
         <artifactId>byte-buddy</artifactId>
-        <version>1.10.6</version>
+        <version>${bytebuddy.version}</version>
       </dependency>
       
       <!-- Metrics -->
